reorganize examples
